Can be merged
Fixed the rest of the initial refactor so that it functions correctly
